Encapsulate associating PG type OIDs with Postgrex extensions in OidDatabase
